Web/requirements.txt
# Name                    Version                   Build  Channel
argon2-cffi               20.1.0                   pypi_0    pypi
async-generator           1.10                     pypi_0    pypi
attrs                     20.3.0                   pypi_0    pypi
backcall                  0.2.0                    pypi_0    pypi
bleach                    3.2.1                    pypi_0    pypi
certifi                   2020.6.20          pyhd3eb1b0_3
cffi                      1.14.3                   pypi_0    pypi
click                     7.1.2                    pypi_0    pypi
cmake                     3.18.4.post1             pypi_0    pypi
colorama                  0.4.4                    pypi_0    pypi
decorator                 4.4.2                    pypi_0    pypi
defusedxml                0.6.0                    pypi_0    pypi
dlib                      19.21.0                  pypi_0    pypi
dnspython                 2.1.0                    pypi_0    pypi
entrypoints               0.3                      pypi_0    pypi
face-recognition          1.3.0                    pypi_0    pypi
face-recognition-models   0.3.0                    pypi_0    pypi
flask                     1.1.2                    pypi_0    pypi
importlib-metadata        3.0.0                    pypi_0    pypi
ipykernel                 5.3.4                    pypi_0    pypi
ipython                   7.16.1                   pypi_0    pypi
ipython-genutils          0.2.0                    pypi_0    pypi
ipywidgets                7.5.1                    pypi_0    pypi
itsdangerous              1.1.0                    pypi_0    pypi
jedi                      0.17.2                   pypi_0    pypi
jinja2                    2.11.2                   pypi_0    pypi
jsonschema                3.2.0                    pypi_0    pypi
jupyter                   1.0.0                    pypi_0    pypi
jupyter-client            6.1.7                    pypi_0    pypi
jupyter-console           6.2.0                    pypi_0    pypi
jupyter-core              4.7.0                    pypi_0    pypi
jupyterlab-pygments       0.1.2                    pypi_0    pypi
markupsafe                1.1.1                    pypi_0    pypi
mistune                   0.8.4                    pypi_0    pypi
nbclient                  0.5.1                    pypi_0    pypi
nbconvert                 6.0.7                    pypi_0    pypi
nbformat                  5.0.8                    pypi_0    pypi
nest-asyncio              1.4.3                    pypi_0    pypi
notebook                  6.1.5                    pypi_0    pypi
numpy                     1.19.3                   pypi_0    pypi
packaging                 20.4                     pypi_0    pypi
pandocfilters             1.4.3                    pypi_0    pypi
parso                     0.7.1                    pypi_0    pypi
pickleshare               0.7.5                    pypi_0    pypi
pillow                    8.0.1                    pypi_0    pypi
pip                       20.2.4           py36haa95532_0
prometheus-client         0.9.0                    pypi_0    pypi
prompt-toolkit            3.0.8                    pypi_0    pypi
pycparser                 2.20                     pypi_0    pypi
pygments                  2.7.2                    pypi_0    pypi
pymongo                   3.11.4                   pypi_0    pypi
pyparsing                 2.4.7                    pypi_0    pypi
pyrsistent                0.17.3                   pypi_0    pypi
python                    3.6.9                h5500b2f_0
python-dateutil           2.8.1                    pypi_0    pypi
pywin32                   300                      pypi_0    pypi
pywinpty                  0.5.7                    pypi_0    pypi
pyzmq                     20.0.0                   pypi_0    pypi
qtconsole                 4.7.7                    pypi_0    pypi
qtpy                      1.9.0                    pypi_0    pypi
send2trash                1.5.0                    pypi_0    pypi
setuptools                50.3.1           py36haa95532_1
six                       1.15.0                   pypi_0    pypi
sqlite                    3.33.0               h2a8f88b_0
terminado                 0.9.1                    pypi_0    pypi
testpath                  0.4.4                    pypi_0    pypi
tornado                   6.1                      pypi_0    pypi
traitlets                 4.3.3                    pypi_0    pypi
vc                        14.1                 h0510ff6_4
vs2015_runtime            14.16.27012          hf0eaf9b_3
wcwidth                   0.2.5                    pypi_0    pypi
webencodings              0.5.1                    pypi_0    pypi
werkzeug                  1.0.1                    pypi_0    pypi
wheel                     0.35.1             pyhd3eb1b0_0
widgetsnbextension        3.5.1                    pypi_0    pypi
wincertstore              0.2              py36h7fe50ca_0
zipp                      3.4.0                    pypi_0    pypi
zlib                      1.2.11               h62dcd97_4